Fix seq_num MsgId::gen() bug
This bugfix fixes an unexpected behaviour where the MsgId for a keyload message on c86/PC differs from the MsgId on EPS32-C3. Values of type usize should not be used to create a MsgId because usize has a platform specific byte size which results in different MsgId values depending on the platform architecture (32- or 64-bit address bus).

+++ lets/src/address.rs
@@ -245,7 +245,7 @@ impl MsgId {
         Self(bytes)
     }
 
-    pub fn gen(appaddr: AppAddr, identifier: &Identifier, topic: &Topic, seq_num: usize) -> MsgId {
+    pub fn gen(appaddr: AppAddr, identifier: &Identifier, topic: &Topic, seq_num: u32) -> MsgId {
         let mut s = Spongos::<KeccakF1600>::init();
         s.absorb(appaddr);
         s.absorb(identifier);
